Conflicts:
	src/components/public/app/admin/email/autoresponder/autoresponder.html

+++ src/core/ang/directives/smDelete.js
@@ -8,7 +8,11 @@ app.directive( 'smDelete', function()
 					.modal({
 						allowMultiple: true,
 						onApprove: function(){
-							scope.deleteResource(attributes.smDelete);
+                            if( attributes.deleteFunction && typeof scope[ attributes.deleteFunction ] == 'function' )
+                                scope[ attributes.deleteFunction ]( attributes.smDelete );
+                            else
+							    scope.deleteResource(attributes.smDelete);
+							$(".small.delete.modal").modal('hide');
 							return true;
 						},
 						onHidden: function() {
